+++ src/components/LoginModal.js
@@ -4,7 +4,7 @@ import Button from "react-bootstrap/Button";
 import { motion } from "framer-motion";
 
 import { useState } from "react";
-import { useLogin } from "../utils/useLogin";
+import { useLogin } from "../hooks/useLogin";
 import { useNavigate  } from "react-router-dom"
 
 import "../styles/Button.css";
