--- src/utils/useBookAndListingContext.js
@@ -1,12 +1,0 @@
-import { BookAndListingContext } from "../contexts/BookAndListingContext";
-import { useContext } from "react";
-
-export const useBookAndListingContext = () => {
-    const context = useContext(BookAndListingContext)
-
-    if (!context){
-        throw Error('useBookAndListingContext must be used inside an BookAndListingContextProvider')
-    }
-
-    return context
-}
